requirements.txt
Flask==2.3.2
Flask_RESTful==0.3.9
PyYAML==6.0
Requests==2.30.0
SQLAlchemy==2.0.12
Werkzeug==2.3.3
psycopg2==2.9.6
